refactor: rename utils/http.ts to utils/server.ts
