+++ test/tests/integration/middleware/cors.test.ts
@@ -2,7 +2,7 @@ import type {Server} from 'node:http';
 import request, {type Response} from 'supertest';
 import {expect} from 'chai';
 
-import {getTestServer} from '../../../utils/http.js';
+import {getTestServer} from '../../../utils/server.js';
 
 describe('cors', function () {
 	this.timeout(15_000);
