+++ test/tests/integration/middleware/etag.test.ts
@@ -2,7 +2,7 @@ import type {Server} from 'node:http';
 import request, {type Response} from 'supertest';
 import {expect} from 'chai';
 
-import {getTestServer} from '../../../utils/http.js';
+import {getTestServer} from '../../../utils/server.js';
 
 describe('etag', function () {
 	this.timeout(15_000);
