+++ test/tests/integration/probes/get-probes.test.ts
@@ -17,7 +17,7 @@ describe('Get Probes', function () {
 
 	before(async () => {
 		await td.replaceEsm('../../../../src/lib/cache/redis-cache.ts', {}, RedisCacheMock);
-		const http = await import('../../../utils/http.js');
+		const http = await import('../../../utils/server.js');
 		addFakeProbe = http.addFakeProbe;
 		deleteFakeProbe = http.deleteFakeProbe;
 		const app = await http.getTestServer();
